==> 02_Algorithms_0.1.1.ipynb <==
# 1+p Log Transformation To Target and Binary Encoding (if required)

import os

import numpy as np
import pandas as pd

import joblib

from category_encoders import BinaryEncoder
from sklearn.preprocessing import FunctionTransformer

df_all = pd.read_pickle('data/df/df_all.pkl')

continuous_features = joblib.load('data/iterables/continuous_features.joblib')
categorical_features = joblib.load('data/iterables/categorical_features.joblib')
target_features = joblib.load('data/iterables/target_features.joblib')

### Execute below if you need data for algorithms that don't support categorical variables
### ============================================================================

be = BinaryEncoder(cols=categorical_features, return_df=True, handle_unknown='ignore')
df_encoded = be.fit_transform(df_all[categorical_features])

joblib.dump(be, 'models/preprocessing/be.joblib')

df_encoded.head()

categorical_features_encoded = df_encoded.columns.tolist()
print(categorical_features_encoded.__len__())

joblib.dump(categorical_features_encoded, 'data/iterables/categorical_features_encoded.joblib')

df_all.reset_index(inplace=True, drop=True)
df_encoded.reset_index(inplace=True, drop=True)

df_all = df_all.merge(df_encoded, left_index=True, right_index=True, how='inner')

### ============================================================================

df_train_dataset = df_all[df_all['source'] == 'train'].copy()
df_train_dataset.reset_index(inplace=True, drop=True)

df_validation_dataset = df_all[df_all['source'] == 'validation'].copy()
df_validation_dataset.reset_index(inplace=True, drop=True)

df_train_dataset.shape, df_validation_dataset.shape

target_transformer = FunctionTransformer(func=np.log1p, inverse_func=np.expm1, validate=True)
df_train_dataset[target_features] = target_transformer.transform(df_train_dataset[target_features].values)

df_train_dataset.to_pickle('data/df/df_train_dataset.pkl')
df_validation_dataset.to_pickle('data/df/df_validation_dataset.pkl')

joblib.dump(target_transformer, 'models/preprocessing/target_transformer.joblib')
