==> 03_Hyperparameters_0.1.ipynb <==
# LGBM Hyperparameter Optimization

import os

import numpy as np
import pandas as pd

import joblib

import lightgbm as lgb

from tqdm import tqdm_notebook
import scipy.optimize as spo

from sklearn.metrics import mean_squared_error
from sklearn.model_selection import train_test_split
from sklearn.model_selection import KFold
from sklearn.preprocessing import FunctionTransformer

from hyperopt import hp, fmin, tpe, STATUS_OK, Trials

def run_lgbm(X_train, X_test, y_train, y_test, feature_names, categorical_features='auto', model_params=None, fit_params=None, seed=21):

    X_train_GBM = lgb.Dataset(X_train, label=y_train, feature_name=feature_names, categorical_feature=categorical_features, free_raw_data=False)
    X_test_GBM = lgb.Dataset(X_test, label=y_test, reference=X_train_GBM, feature_name=feature_names, free_raw_data=False)
    
    if model_params is None:
        model_params = {'seed': seed, 'num_threads': 16, 'objective':'root_mean_squared_error', 
                        'metric': ['root_mean_squared_error'] }
        
    if fit_params is None:
        fit_params = {'verbose_eval': True, 'num_boost_round': 300, 'valid_sets': [X_test_GBM], 
                      'early_stopping_rounds': 30, 'categorical_feature': categorical_features, 'feature_name': feature_names}
        
    model = lgb.train(model_params, X_train_GBM, **fit_params)
    y_pred = model.predict(X_test, model.best_iteration)
    return model, y_pred, mean_squared_error(y_test, y_pred)

## Lgbm

df_train_dataset = pd.read_pickle('data/df/df_train_dataset.pkl')
df_validation_dataset = pd.read_pickle('data/df/df_validation_dataset.pkl')

continuous_features = joblib.load('data/iterables/continuous_features.joblib')
categorical_features = joblib.load('data/iterables/categorical_features.joblib')
target_features = joblib.load('data/iterables/target_features.joblib')
target_transformer = joblib.load('models/preprocessing/target_transformer.joblib')

df_train_dataset.shape, df_validation_dataset.shape

X = df_train_dataset[categorical_features + continuous_features].values
y = df_train_dataset[target_features].values.flatten()
print(X.shape, y.shape)

X_train, X_test, y_train, y_test = train_test_split(X, y, train_size=0.85, test_size=0.15, shuffle=True, random_state=10)
print(X_train.shape, X_test.shape, y_train.shape, y_test.shape)

feature_names = categorical_features + continuous_features
X_train_GBM = lgb.Dataset(X_train, label=y_train, feature_name=feature_names, categorical_feature=categorical_features, free_raw_data=False)
X_test_GBM = lgb.Dataset(X_test, label=y_test, reference=X_train_GBM, feature_name=feature_names, free_raw_data=False)

space = {
    'learning_rate': hp.uniform('learning_rate', 0.001, 0.04),
    'max_depth' : hp.quniform("max_depth", 3, 8, 1),
    'bagging_fraction': hp.uniform('bagging_fraction', 0.85, 1.0),  
    'bagging_freq': hp.quniform('bagging_freq', 1, 16, 1),
    'lambda_l1': hp.uniform('lambda_l1', 4.0, 16.0), 
    'lambda_l2': hp.uniform('lambda_l2', 4.0, 16.0),
    'feature_fraction': hp.uniform('feature_fraction', 0.10, 0.30),
    'num_leaves': hp.quniform("num_leaves", 4, 64, 1),
    'min_data_in_leaf': hp.quniform("min_data_in_leaf", 32, 512, 1),
    'max_delta_step': hp.uniform('max_delta_step', 0.0, 0.5),
    'min_data_per_group': hp.quniform('min_data_per_group', 16, 512, 1),
    'min_sum_hessian_in_leaf' : hp.uniform('min_sum_hessian_in_leaf', 0.5, 4.0),
    'min_gain_to_split': hp.uniform('min_gain_to_split', 0.0, 1.0),
    'max_bin' : hp.quniform("max_bin", 64, 512, 1),
    'min_data_in_bin': hp.quniform("min_data_in_bin", 1, 32, 1),
    'max_cat_threshold': hp.quniform('max_cat_threshold', 16, 64, 1),
    'cat_l2' : hp.uniform('cat_l2', 32.0, 128.0),
    'cat_smooth': hp.uniform('cat_smooth', 0.0, 4.0),
    'sparse_threshold' : hp.uniform('sparse_threshold', 0.16, 0.64),

}

param_ints = ['max_depth', 'bagging_freq', 'num_leaves', 'min_data_in_leaf', 'min_data_per_group', 'max_bin', 
              'min_data_in_bin', 'max_cat_threshold']
def objective(space):
    model_params = {'num_threads': 16, 'objective': 'root_mean_squared_error', 
                    'metric': ['root_mean_squared_error'], 'boosting': 'gbdt'}
    
    for param in space.keys():
        if param in param_ints:
            model_params[param] = int(space[param])
        else:
            model_params[param] = space[param]
        
    fit_params = {'verbose_eval': False, 'num_boost_round': 300, 'valid_sets': [X_test_GBM], 'early_stopping_rounds': 30, 
                 'categorical_feature': categorical_features, 'feature_name': feature_names}
    
    model = lgb.train(model_params, X_train_GBM, **fit_params)  
    y_pred = model.predict(X_test, model.best_iteration)
    
    results = {'loss': mean_squared_error(y_test, y_pred) ** .5, 'status': STATUS_OK }
    for param in space.keys():
        results[param] = space[param]
    return results

trials = Trials()
best_parameters = fmin(fn=objective, space=space, algo=tpe.suggest, max_evals=100, trials=trials)

print(best_parameters)
joblib.dump(best_parameters, 'models/parameters/best_parameters2.joblib')

df = pd.DataFrame(trials.results)
df.nsmallest(10, 'loss')
